Split DbReader reference searches for #2115
