+++ changelog.d/unreleased/2115.internal.md
@@ -1,0 +1,16 @@
+---
+category: internal
+issues:
+  - 2115
+affected:
+  - src/CodeIndex/Database/DbReader.cs
+  - src/CodeIndex/Database/DbReader.References.cs
+---
+
+## English
+
+- **Split reference search readers out of `DbReader.cs` (#2115)** — reference lookup and count helpers now live in their own partial `DbReader` file, reducing the remaining reader monolith without changing query behavior.
+
+## 日本語
+
+- **参照検索 reader を `DbReader.cs` から分割しました (#2115)** — 参照検索と件数取得のヘルパーを専用の partial `DbReader` ファイルへ移し、クエリ挙動を変えずに reader 本体を小さくしました。
